Refactor generate features integration tests
